sovereign-agent-v0.2.1/sovereign-agent/pyproject.toml
[project]
name = "sovereign-agent"
version = "0.2.3"
description = "Sovereign local agent on Ollama. FOSS-first. Authority-tiered. Yours."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "operator" }]

dependencies = [
    "pydantic>=2.6,<3",
    "ollama>=0.4.0",
    "structlog>=24.1",
    "httpx>=0.27",
    "typer>=0.12",
    "python-ulid>=2.7",
    "tiktoken>=0.7",
    "sqlite-vec>=0.1.6",
    "pyyaml>=6.0",
    "rich>=13.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "hypothesis>=6.100",
    "ruff>=0.5",
    "mypy>=1.10",
]
distill = [
    "pymupdf4llm>=0.0.17",
    "tree-sitter>=0.22",
    "tree-sitter-languages>=1.10",
]
media = [
    # Ported from mos_harvester / mos_merger — heavy GPU deps, optional
    "faster-whisper>=1.0",
    "easyocr>=1.7",
    "PyMuPDF>=1.24",
]

[project.scripts]
sovereign = "sovereign_agent.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/sovereign_agent"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-q --strict-markers"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "S", "RET", "SIM", "PTH"]
ignore = ["S101"]  # assert in tests is fine

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
